Feature/product optimize

+++ src/main/java/com/example/team5_project/dto/search/response/FindSearchResponse.java
@@ -1,4 +1,4 @@
 package com.example.team5_project.dto.search.response;
 
-public record FindSearchResponse(String name, Long count) {
+public record FindSearchResponse(String name, int count) {
 }
